Move `HypreStealOwnership` in `hypre.hpp/cpp`.
